refactor: split SecurityControlHasACallerTests into partials by guard family (#643)
SecurityControlHasACallerTests.cs had grown to 1265 lines, well over this
repo's file-size hard cap. This is a pure mechanical reorganization
following the repo's documented Partial Class Pattern (e.g.
PlanExecutor.Scheduling.cs, PlanExecutor.Recovery.cs) — no behavior change,
no logic edits, every member moved verbatim. The class is now split across
five files: SecurityControlHasACallerTests.cs (usings, class-level docs,
governance-contract and per-run-fact guards), .ValidatorWiring.cs (the
validator-invocation-mechanism and MediatR-behavior guards),
.DeclaringFilesAndDeadValidators.cs (namespace-collision controls and the
dead-validator guard), .SkillEgress.cs (the skill-egress consumer-resolved
guard), and .Helpers.cs (the shared private static parsing helpers). All 12
tests in the class pass identically before and after, and the full
Presentation.Common.Tests suite (303 tests) passes clean.
